Updated return type of some functions.

+++ echo.h
@@ -1,5 +1,5 @@
 struct echo_cmd
 {
     char data[1024];
-    unsigned int bytes_remaining;
+    unsigned long bytes_remaining;
 };
